Add org lists on org index and manage signators page
Add new widgets to the index page of the organizations app. Create a
table of my organizations, with action buttons for editing, swapping
signators and deleting. This will eliminate the need for a separate page
and reduce the number of clicks.

In addition, make another table on the index page with a list of all
organizations. The list is acquired by AJAX, and has a search field
that can filter without reloading the page.

In order to faciliate AJAX request, an `api.py` file was created in the
views. This API file will handle all json api requests, to which I may
add for use in other apps.

Started a manage signators page, which currently lists all signators.
Consider later creating an API to filter and search signators.